refactor: extract game screen fallback next-turn into part file (#3878)
Move the Flame-canvas fallback Next-turn button and turn-resolution
orchestration from game_screen.dart into a library part for Phase 3
modularization without changing public API or overlay behavior.
